requirements.txt
# Core dependencies
torch>=1.9.0
torchvision>=0.10.0
timm>=0.4.12
ml_collections>=0.1.1
numpy>=1.19.2
Pillow>=8.3.1
scikit-image>=0.18.3
scikit-learn>=0.24.2
tqdm>=4.62.3

# Diffusion model dependencies
accelerate==0.12.0
absl-py>=0.13.0
einops>=0.4.1
ftfy==6.1.1
transformers==4.23.1

# Image processing
opencv-python>=4.5.3.56
albumentations>=1.1.0

# Utilities
matplotlib>=3.4.3
pandas>=1.3.3
wandb>=0.12.0
tensorboard>=2.7.0

# Optional: for development
pytest>=6.2.5
black>=21.9b0
flake8>=3.9.2
isort>=5.9.3